Fix Twofish Encryption
There were a couple problems with the previous implementation.

A long than 32 character encryption password caused errors. This was
solved by using a sha384 hash of the password as the password which has
a 32 character digest.

Previously if decryption failed the program threw an exception. This is
now properly encoded into a failure reason for reporting to the user and
gracefully exiting.

An enhancement was also added to allow viewing the initialization
vector in the print usage command.