Add Challenge, Cleanup Failures, Sort Files
Adds an encrypted challenge to the device when it is being formatted,
this is an unknown secret text encrypted with the IV and user's master
password. This allows us to verify the user is using the correct master
password without having them enter it twice when adding new things to
the device. If the challenge fails to decrypt and verify correctly, then
we know they entered an incorrect password.

Cleanup failure reasons. Finally splits this into multiple types which
can be recombined into a single type using the failure reason
translator.

Sorts files by name when listing them instead of by file id.