Enable Full Encryption
Previously only the passwords were encrypted, this enables encryption
for all data stored in the device (with the exception of inode headers
and length fields).

While this does not prevent someone from removing files or listing file
ids, no account information can be leaked off of the device without the
password.