Reset Hardware Before Transmissions
Previously the client would wait for the hardware to send a ready signal
before attempting any communications. This previously relied on the fact
that connecting with boost asio would toggle the DTR line. When the
hardware was commited to use the CH330N chip it removed the DTR line but
left available a RTS line. This change toggles the RTS line after
connect to reset the device.