Add Rename File Command
Adds ability to rename a file without having to manually
read/create/delete. This is purely client side driven and requires no
API changes. The new file is created before the old is delete to reduce
the chance of loosing data; additionally the old file is not deleted if
the new file is not created