Cleanup of Arguments
Cleans up the usage and naming of arguments to make the flags more
intuitive. Generally they follow the rule lower case when adding and
upper case when modifying. Uniforms terminology around specifying a file
as a parameter to always use `filename` when only one is expected, and
only get more specific when multiples are needed (e.g. `rename_from` and
`rename_to`)