Decouple UI and Business Logic
Previously the user interactions (text input / output) were embedded in
the logic that was implementing the high-level actions. This was a SRP
violation and not conducive to adding different UIs. This commit
refactors the `commands/` folder to remove all user interaction code,
and adds a separate `ui/` folder to hold all user interaction code.